+++ changelog.md
@@ -27,6 +27,7 @@
 - Converted some tests for old pages into the new liveview pages
 - Moved nearly all of the `Central` stuff over to be part of Teiserver
 - Started making progress on
+- Added `hms_or_hms_ymd` format to `TimexHelper`
 
 ## v1.1.1